@@ -34,11 +34,12 @@
     "babel-preset-es2015": "^6.0.15",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.0.15",
+    "css-loader": "^0.23.1",
     "eslint": "^1.10.3",
     "eslint-plugin-react": "^3.6.2",
     "html-loader": "^0.4.3",
     "react-hot-loader": "^1.3.0",
-    "webpack": "^1.12.2",
+    "webpack": "^1.13.2",
     "webpack-dev-server": "^1.12.1"
   },
   "dependencies": {
@@ -47,12 +48,12 @@
     "browserify-versionify": "^1.0.6",
     "extend": "^3.0.0",
     "ify-loader": "^1.0.3",
+    "json-loader": "^0.5.4",
     "jzz": "^0.3.6",
     "material-ui": "^0.15.3",
     "pixi.js": "^4.0.0-rc3",
     "react": "^15.2.1",
-    "react-addons": "^0.9.1-deprecated",
-    "react-bootstrap": "^0.30.2",
+    "react-bootstrap": "^0.30.3",
     "react-dom": "^0.14.6",
     "react-hotkey": "^0.6.0",
     "react-tap-event-plugin": "^1.0.0",
