@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import update from 'react-addons-update';
 import PIXI from 'pixi.js';
 import TinyColor from 'tinycolor2';
-import hotkey from 'react-hotkey';
 import { StyleSheet, css } from 'aphrodite';
 
 export default class Piano extends Component {
@@ -10,13 +9,8 @@ export default class Piano extends Component {
     super(props, context);
 
     this.state = {
-      keys: [],
-      addUnison: false
+      keys: []
     }
-
-    hotkey.activate('keydown');
-    hotkey.activate('keyup');
-    this.hotkeyHandler = this.handleHotkey.bind(this);
 
     const res = this.res = 2; // renderer resolution
     const height = this.h = props.height / this.res; // renderer height adjusted for resolution boost
@@ -26,7 +20,6 @@ export default class Piano extends Component {
 
   componentDidMount() {
     window.addEventListener('resize', this.handleResize.bind(this));
-    hotkey.addHandler(this.hotkeyHandler);
 
     const width = window.innerWidth / this.res;
 
@@ -38,12 +31,11 @@ export default class Piano extends Component {
     this.stage.interactive = true;
     this.animate();
 
-    this.generatePiano(this.props.noteStack);
+    this.generatePiano(this.props.notes);
   }
 
   componentWillUnmount() {
     window.removeEventListener('resize', this.handleResize);
-    hotkey.removeHandler(this.hotkeyHandler);
   }
 
   componentWillUpdate() {
@@ -51,21 +43,16 @@ export default class Piano extends Component {
   }
 
   componentWillReceiveProps(nextProps) {
-  	this.generatePiano(nextProps.noteStack)
-  }
-
-  handleHotkey(e) {
-    // if option-key is down, add unisons
-    if (e.which === 18)
-        this.setState({ addUnison: e.type === 'keydown' ? true : false })
+  	this.generatePiano(nextProps.notes)
   }
 
   handleResize() {
-    this.renderer.view.style.width = window.innerWidth + 'px';
+    this.renderer.view.style.width = window.innerWidth;
+    this.refs.pianoCanvas.width = window.innerWidth;
     this.renderPiano();
   }
 
-  generatePiano(noteStack) {
+  generatePiano(notes) {
     var keys = new Array(128); // number of possible midi notes
     let numVisibleKeys = this.props.octaves * 12;
     let lowerBound = this.props.midiBase;
@@ -75,7 +62,7 @@ export default class Piano extends Component {
     for (var i = 0; i < keys.length; i++) {
     	// count occurrences of key in current note stack
     	let count = 0;
-			for(var note of noteStack)
+			for(var note of notes)
 			    if(note === i) count++;
 
       keys[i] = {
@@ -168,7 +155,7 @@ export default class Piano extends Component {
 
     if (key.unisons === -1)
       this.props.onClick(key, true)
-    else if (this.state.addUnison) {
+    else if (this.props.addUnison) {
       if (key.unisons < 4)
         this.props.onClick(key, true)
     }
@@ -193,7 +180,7 @@ Piano.propTypes = {
   midiBase: React.PropTypes.number.isRequired,
   height: React.PropTypes.number.isRequired,
   accentColor: React.PropTypes.string.isRequired,
-  noteStack: React.PropTypes.array.isRequired,
+  notes: React.PropTypes.array.isRequired,
   onClick: React.PropTypes.func.isRequired,
 }
 
@@ -202,7 +189,6 @@ const styles = StyleSheet.create({
     piano: {
       'position': 'fixed',
       'bottom': -5,
-      'right': 0,
       'padding': 'none',
       'margin': 'none'
     }
