@@ -1,47 +1,47 @@
-export const chordTypes = [
-  { notes: [0], name: "single note", color: 0x364C87 }, // dark blue
-  { notes: [0, 0], name: "double unison", color: 0x364C87 }, // dark blue
-  { notes: [0, 0, 0], name: "triple unison", color: 0x364C87 }, // dark blue
-  { notes: [0, 0, 0, 0], name: "quadruple unison", color: 0x364C87 }, // dark blue
-  { notes: [0, 1], name: "semitone", color: 0x3A859E }, // teil
-  { notes: [0, 2], name: "major second", color: 0x66D4AD }, // light green
-  { notes: [0, 3], name: "minor third", color: 0x30ED0E }, // lime green
-  { notes: [0, 4], name: "major third", color: 0xF2F211 }, // yellow 
-  { notes: [0, 5], name: "perfect fourth", color: 0xF5B92F }, // orange
-  { notes: [0, 6], name: "tritone", color: 0xE83838 }, // red
-  { notes: [0, 7], name: "perfect fifth", color: 0xF5B92F }, // orange
-  { notes: [0, 8], name: "minor sixth", color: 0xF2F211 }, // yellow
-  { notes: [0, 9], name: "major sixth", color: 0x30ED0E }, // lime green
-  { notes: [0, 10], name: "minor seventh", color: 0x66D4AD }, // light green
-  { notes: [0, 11], name: "major seventh", color: 0x3A859E }, // teil
-  { notes: [0, 1, 2], name: "chromatic cluster", color: 0xE6B89C },
-  { notes: [0, 3, 6], name: "diminished triad", color: 0xE6B89C },
-  { notes: [0, 3, 7], name: "minor triad", color: 0xE6B89C },
-  { notes: [0, 4, 7], name: "major triad", color: 0xE6B89C },
-  { notes: [0, 2, 7], name: "sus triad", color: 0xE6B89C },
-  { notes: [0, 4, 8], name: "augmented triad", color: 0xE6B89C },
-  { notes: [0, 3, 8], name: "major 6/3 triad", color: 0xE6B89C },
-  { notes: [0, 4, 9], name: "minor 6/3 triad", color: 0xE6B89C },
-  { notes: [0, 5, 9], name: "major 6/4 triad", color: 0xE6B89C },
-  { notes: [0, 5, 8], name: "minor 6/4 triad", color: 0xE6B89C },
-  { notes: [0, 3, 6, 9], name: "diminished-seventh", color: 0xE6B89C }, 
-  { notes: [0, 2, 6, 8], name: "french sixth", color: 0xE6B89C },
-  { notes: [0, 2, 5, 7], name: "suspended seventh", color: 0xE6B89C }, 
-  { notes: [0, 4, 7, 10], name: "dominant seventh", color: 0xE6B89C },
-  { notes: [0, 2, 5, 8], name: "half-diminished", color: 0xE6B89C },
-  { notes: [0, 1, 5, 8], name: "major-seventh", color: 0xE6B89C },
-  { notes: [0, 3, 5, 8], name: "minor-seventh", color: 0xE6B89C },
-  { notes: [0, 1, 4, 8], name: "min/maj-seventh", color: 0xE6B89C },
-  { notes: [0, 3, 6, 8], name: "dominant seventh 6/5", color: 0xE6B89C },
-  { notes: [0, 3, 5, 9], name: "dominant seventh 4/3", color: 0xE6B89C },
-  { notes: [0, 2, 6, 9], name: "dominant seventh 4/2", color: 0xE6B89C },
-  { notes: [0, 2, 4, 7, 9], name: "pentatonic", color: 0xE6B89C },
-  { notes: [0, 2, 4, 6, 9], name: "dominant-ninth", color: 0xE6B89C },
-  { notes: [0, 2, 4, 6, 8, 10], name: "whole-tone", color: 0xE6B89C },
-  { notes: [0, 1, 3, 5, 6, 8, 10], name: "diatonic", color: 0xE6B89C },
-  { notes: [0, 1, 3, 4, 6, 8, 10], name: "acoustic", color: 0xE6B89C },
-  { notes: [0, 1, 3, 4, 6, 7, 9, 10], name: "octatonic", color: 0xE6B89C },
-  { notes: [0, 1, 3, 5, 6, 8, 9], name: "harmonic major", color: 0xE6B89C },
-  { notes: [0, 1, 3, 4, 6, 8, 9], name: "harmonic minor", color: 0xE6B89C },
-  { notes: [0, 1, 4, 5, 8, 9], name: "hexatonic", color: 0xE6B89C }
+export const chords = [
+  { notes: [0], name: "single note"},
+  { notes: [0, 0], name: "double unison"},
+  { notes: [0, 0, 0], name: "triple unison"},
+  { notes: [0, 0, 0, 0], name: "quadruple unison"},
+  { notes: [0, 1], name: "semitone"},
+  { notes: [0, 2], name: "major second"},
+  { notes: [0, 3], name: "minor third"},
+  { notes: [0, 4], name: "major third"},
+  { notes: [0, 5], name: "perfect fourth"},
+  { notes: [0, 6], name: "tritone"},
+  { notes: [0, 7], name: "perfect fifth"},
+  { notes: [0, 8], name: "minor sixth"},
+  { notes: [0, 9], name: "major sixth"},
+  { notes: [0, 10], name: "minor seventh"},
+  { notes: [0, 11], name: "major seventh"},
+  { notes: [0, 1, 2], name: "chromatic cluster"},
+  { notes: [0, 3, 6], name: "diminished triad"},
+  { notes: [0, 3, 7], name: "minor triad"},
+  { notes: [0, 4, 7], name: "major triad"},
+  { notes: [0, 2, 7], name: "sus triad"},
+  { notes: [0, 4, 8], name: "augmented triad"},
+  { notes: [0, 3, 8], name: "major 6/3 triad"},
+  { notes: [0, 4, 9], name: "minor 6/3 triad"},
+  { notes: [0, 5, 9], name: "major 6/4 triad"},
+  { notes: [0, 5, 8], name: "minor 6/4 triad"},
+  { notes: [0, 3, 6, 9], name: "diminished-seventh"}, 
+  { notes: [0, 2, 6, 8], name: "french sixth"},
+  { notes: [0, 2, 5, 7], name: "suspended seventh"}, 
+  { notes: [0, 4, 7, 10], name: "dominant seventh"},
+  { notes: [0, 2, 5, 8], name: "half-diminished"},
+  { notes: [0, 1, 5, 8], name: "major-seventh"},
+  { notes: [0, 3, 5, 8], name: "minor-seventh"},
+  { notes: [0, 1, 4, 8], name: "min/maj-seventh"},
+  { notes: [0, 3, 6, 8], name: "dominant seventh 6/5"},
+  { notes: [0, 3, 5, 9], name: "dominant seventh 4/3"},
+  { notes: [0, 2, 6, 9], name: "dominant seventh 4/2"},
+  { notes: [0, 2, 4, 7, 9], name: "pentatonic"},
+  { notes: [0, 2, 4, 6, 9], name: "dominant-ninth"},
+  { notes: [0, 2, 4, 6, 8, 10], name: "whole-tone"},
+  { notes: [0, 1, 3, 5, 6, 8, 10], name: "diatonic"},
+  { notes: [0, 1, 3, 4, 6, 8, 10], name: "acoustic"},
+  { notes: [0, 1, 3, 4, 6, 7, 9, 10], name: "octatonic"},
+  { notes: [0, 1, 3, 5, 6, 8, 9], name: "harmonic major"},
+  { notes: [0, 1, 3, 4, 6, 8, 9], name: "harmonic minor"},
+  { notes: [0, 1, 4, 5, 8, 9], name: "hexatonic"}
 ]
